requirements.txt
asgiref==3.8.1
attrs==25.3.0
auth_lib_profcomff==2024.8.4
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.4.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg==3.2.6
psycopg-binary==3.2.6
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
